requirements.txt
rumps
pyobjc-framework-Quartz
sounddevice
soundfile
numpy
pydub
openai
